Fixed tests!
