+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "python.testing.pytestArgs": [
+        "."
+    ],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true,
+    "python.pythonPath": "./.venv/bin/python",
+    "python.defaultInterpreterPath": "./venv/bin/python"
+}
